Adds links to contact items
